% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ver_1.0.R
\name{match_protein_interaction}
\alias{match_protein_interaction}
\title{Title Get the list of interators for proteins base on the BioGrid database}
\usage{
match_protein_interaction(list)
}
\arguments{
\item{list}{an input list containing proteins to be analyzed}
}
\value{
an output list containing the original proteins with their interactors
}
\description{
Title Get the list of interators for proteins base on the BioGrid database
}
\examples{
match_protein_interaction(pro_list)
}
