% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ver_1.0.R
\name{matrix_boxing}
\alias{matrix_boxing}
\title{Quickly boxing continuous matrix}
\usage{
matrix_boxing(m, levels)
}
\arguments{
\item{m}{an input matrix}

\item{levels}{number of categories to be cut, depending on the quantiles (using ::cut())}
}
\value{
the output matrix
}
\description{
Quickly boxing continuous matrix
}
\examples{
data_boxing(matrix, 2)
}
