% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ver_1.0.R
\name{matrix_pairwise_interation}
\alias{matrix_pairwise_interation}
\title{Title Quickly obtain pairwise interaction matrix}
\usage{
matrix_pairwise_interation(df)
}
\arguments{
\item{df}{an input data.frame object}
}
\value{
interaction matrix
}
\description{
Title Quickly obtain pairwise interaction matrix
}
\examples{
pairwise_interation(df)
}
